feat(user): Adicionando validação de token do usuário

+++ src/services/user.service.ts
@@ -1,4 +1,5 @@
 import type { PrismaClient } from "@prisma/client";
+import jwt, { type JwtPayload } from "jsonwebtoken";
 import bcrypt from "bcrypt";
 import type {
   IUserCreate,
@@ -21,11 +22,18 @@ class UserService {
     return allUsersData;
   }
 
-  async registerNewUser(userInfos: IUserCreate): Promise<IUserPublic> {
+  async registerNewUser(
+    userInfos: IUserCreate,
+    adminUserToken: string,
+  ): Promise<IUserPublic> {
     const saltRounds = process.env.SALT_ROUNDS;
 
     if (!saltRounds) {
       throw new Error("Variável de ambiente SALT_ROUNDS não encontrada.");
+    }
+
+    if (!adminUserToken || !(await this.verifyUserToken(adminUserToken))) {
+      throw new Error("Token inválido.");
     }
 
     if (
@@ -39,7 +47,10 @@ class UserService {
 
     const saltRoundsNumber = parseInt(saltRounds, 10);
 
-    const hashPassword = await bcrypt.hash(userInfos.password, saltRoundsNumber);
+    const hashPassword = await bcrypt.hash(
+      userInfos.password,
+      saltRoundsNumber,
+    );
 
     const newUser: IUserPublic = await this.prisma.user.create({
       data: {
@@ -56,10 +67,14 @@ class UserService {
   async updateUserData(
     userNewData: IUserUpdate,
     userUuid: string,
+    adminToken: string,
   ): Promise<IUserPublic> {
     if (!userUuid) {
       throw new Error(responseMessages.fillAllFieldMessage);
     }
+
+    if (!adminToken || !(await this.verifyUserToken(adminToken)))
+      throw new Error("Token inválido.");
 
     const updateFields = removeUndefinedUpdateFields(userNewData);
 
@@ -73,7 +88,10 @@ class UserService {
     return updatedUser;
   }
 
-  async deleteUserData(userUuid: string): Promise<string> {
+  async deleteUserData(userUuid: string, adminToken: string): Promise<string> {
+    if (!(await this.verifyUserToken(adminToken)))
+      throw new Error("Usuário sem privilégios");
+
     await this.prisma.user.delete({
       where: {
         user_id: userUuid,
@@ -81,6 +99,20 @@ class UserService {
     });
 
     return "Usuário excluido com sucesso";
+  }
+
+  private async verifyUserToken(adminUserToken: string): Promise<boolean> {
+    const JWT_SECRET = process.env.JWT_SECRET;
+
+    if (!JWT_SECRET) throw new Error("Secret inválido.");
+
+    const tokenDecoded = jwt.verify(adminUserToken, JWT_SECRET) as IUserPublic;
+
+    if (tokenDecoded.user_type !== "Admin") {
+      return false;
+    }
+
+    return true;
   }
 }
